requirements.txt
python-telegram-bot==20.7
sentence-transformers
scikit-learn
numpy
